+++ AdventOfCodeConsole/Puzzles/IDay.cs
@@ -2,6 +2,6 @@
 
 public interface IDay
 {
-    int Part1(string input);
-    int Part2(string input);
+    long Part1(string input);
+    long Part2(string input);
 }
